streamlit_app/requirements.txt
# Airflow Providers
apache-airflow-providers-postgres==5.8.0

# Data Processing
pandas==2.0.3
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
streamlit==1.31.0